+++ fractal-gas-web/build-atari/CMakeFiles/fg_tests.dir/tests/test_fractal_gas.cpp.o.d
@@ -210,6 +210,14 @@ CMakeFiles/fg_tests.dir/tests/test_fractal_gas.cpp.o: \
  /home/guillem/fragile/fractal-gas-web/src/env.hpp \
  /home/guillem/fragile/fractal-gas-web/src/kinetic.hpp \
  /home/guillem/fragile/fractal-gas-web/src/swarm_algorithm.hpp \
+ /home/guillem/fragile/fractal-gas-web/src/visit_grid.hpp \
+ /usr/include/c++/15/unordered_map \
+ /usr/include/c++/15/bits/unordered_map.h \
+ /usr/include/c++/15/bits/hashtable.h \
+ /usr/include/c++/15/bits/hashtable_policy.h \
+ /usr/include/c++/15/bits/enable_special_members.h \
+ /usr/include/c++/15/bits/node_handle.h \
+ /usr/include/c++/15/bits/erase_if.h \
  /home/guillem/fragile/fractal-gas-web/src/walker_state.hpp \
  /home/guillem/fragile/fractal-gas-web/tests/mock_env.hpp \
  /usr/include/c++/15/cstring /usr/include/string.h /usr/include/strings.h \
@@ -218,12 +226,6 @@ CMakeFiles/fg_tests.dir/tests/test_fractal_gas.cpp.o: \
  /home/guillem/fragile/fractal-gas-web/src/env.hpp \
  /home/guillem/fragile/fractal-gas-web/tests/test_framework.hpp \
  /usr/include/c++/15/functional /usr/include/c++/15/bits/std_function.h \
- /usr/include/c++/15/unordered_map \
- /usr/include/c++/15/bits/unordered_map.h \
- /usr/include/c++/15/bits/hashtable.h \
- /usr/include/c++/15/bits/hashtable_policy.h \
- /usr/include/c++/15/bits/enable_special_members.h \
- /usr/include/c++/15/bits/node_handle.h \
- /usr/include/c++/15/bits/erase_if.h /usr/include/c++/15/array \
- /usr/include/c++/15/compare \
+ /usr/include/c++/15/array /usr/include/c++/15/compare \
+ /home/guillem/fragile/fractal-gas-web/tests/visit_mock_env.hpp \
  /home/guillem/fragile/fractal-gas-web/tests/fixtures/fixtures_generated.hpp