+++ fractal-gas-web/build-atari/CMakeFiles/fractal_gas_core.dir/src/visit_grid.cpp.o.d
@@ -2,8 +2,7 @@ CMakeFiles/fractal_gas_core.dir/src/visit_grid.cpp.o: \
  /home/guillem/fragile/fractal-gas-web/src/visit_grid.cpp \
  /usr/include/stdc-predef.h \
  /home/guillem/fragile/fractal-gas-web/src/visit_grid.hpp \
- /usr/include/c++/15/array /usr/include/c++/15/compare \
- /usr/include/c++/15/bits/version.h \
+ /usr/include/c++/15/cstdint \
  /usr/include/x86_64-linux-gnu/c++/15/bits/c++config.h \
  /usr/include/x86_64-linux-gnu/c++/15/bits/os_defines.h \
  /usr/include/features.h /usr/include/features-time64.h \
@@ -15,24 +14,6 @@ CMakeFiles/fractal_gas_core.dir/src/visit_grid.cpp.o: \
  /usr/include/x86_64-linux-gnu/gnu/stubs-64.h \
  /usr/include/x86_64-linux-gnu/c++/15/bits/cpu_defines.h \
  /usr/include/c++/15/pstl/pstl_config.h \
- /usr/include/c++/15/initializer_list /usr/include/c++/15/type_traits \
- /usr/include/c++/15/bits/functexcept.h \
- /usr/include/c++/15/bits/exception_defines.h \
- /usr/include/c++/15/bits/stl_algobase.h \
- /usr/include/c++/15/bits/cpp_type_traits.h \
- /usr/include/c++/15/ext/type_traits.h \
- /usr/include/c++/15/ext/numeric_traits.h \
- /usr/include/c++/15/bits/stl_pair.h /usr/include/c++/15/bits/move.h \
- /usr/include/c++/15/bits/utility.h \
- /usr/include/c++/15/bits/stl_iterator_base_types.h \
- /usr/include/c++/15/bits/stl_iterator_base_funcs.h \
- /usr/include/c++/15/bits/concept_check.h \
- /usr/include/c++/15/debug/assertions.h \
- /usr/include/c++/15/bits/stl_iterator.h \
- /usr/include/c++/15/bits/ptr_traits.h /usr/include/c++/15/debug/debug.h \
- /usr/include/c++/15/bits/predefined_ops.h /usr/include/c++/15/bit \
- /usr/include/c++/15/concepts /usr/include/c++/15/bits/range_access.h \
- /usr/include/c++/15/cstdint \
  /usr/lib/gcc/x86_64-linux-gnu/15/include/stdint.h /usr/include/stdint.h \
  /usr/include/x86_64-linux-gnu/bits/libc-header-start.h \
  /usr/include/x86_64-linux-gnu/bits/types.h \
@@ -44,14 +25,31 @@ CMakeFiles/fractal_gas_core.dir/src/visit_grid.cpp.o: \
  /usr/include/x86_64-linux-gnu/bits/stdint-least.h \
  /usr/include/c++/15/unordered_map \
  /usr/include/c++/15/bits/requires_hosted.h \
+ /usr/include/c++/15/initializer_list \
  /usr/include/c++/15/bits/unordered_map.h \
  /usr/include/c++/15/bits/hashtable.h \
  /usr/include/c++/15/bits/hashtable_policy.h /usr/include/c++/15/tuple \
+ /usr/include/c++/15/bits/stl_pair.h /usr/include/c++/15/type_traits \
+ /usr/include/c++/15/bits/version.h /usr/include/c++/15/bits/move.h \
+ /usr/include/c++/15/bits/utility.h \
  /usr/include/c++/15/bits/uses_allocator.h \
  /usr/include/c++/15/bits/invoke.h \
  /usr/include/c++/15/bits/functional_hash.h \
  /usr/include/c++/15/bits/hash_bytes.h \
- /usr/include/c++/15/ext/aligned_buffer.h \
+ /usr/include/c++/15/bits/stl_algobase.h \
+ /usr/include/c++/15/bits/functexcept.h \
+ /usr/include/c++/15/bits/exception_defines.h \
+ /usr/include/c++/15/bits/cpp_type_traits.h \
+ /usr/include/c++/15/ext/type_traits.h \
+ /usr/include/c++/15/ext/numeric_traits.h \
+ /usr/include/c++/15/bits/stl_iterator_base_types.h \
+ /usr/include/c++/15/bits/stl_iterator_base_funcs.h \
+ /usr/include/c++/15/bits/concept_check.h \
+ /usr/include/c++/15/debug/assertions.h \
+ /usr/include/c++/15/bits/stl_iterator.h \
+ /usr/include/c++/15/bits/ptr_traits.h /usr/include/c++/15/debug/debug.h \
+ /usr/include/c++/15/bits/predefined_ops.h /usr/include/c++/15/bit \
+ /usr/include/c++/15/concepts /usr/include/c++/15/ext/aligned_buffer.h \
  /usr/include/c++/15/ext/alloc_traits.h \
  /usr/include/c++/15/bits/alloc_traits.h \
  /usr/include/c++/15/bits/stl_construct.h /usr/include/c++/15/new \
@@ -64,6 +62,7 @@ CMakeFiles/fractal_gas_core.dir/src/visit_grid.cpp.o: \
  /usr/include/c++/15/bits/stl_function.h \
  /usr/include/c++/15/backward/binders.h \
  /usr/include/c++/15/bits/node_handle.h \
+ /usr/include/c++/15/bits/range_access.h \
  /usr/include/c++/15/bits/erase_if.h \
  /usr/include/c++/15/bits/memory_resource.h /usr/include/c++/15/cstddef \
  /usr/lib/gcc/x86_64-linux-gnu/15/include/stddef.h \
@@ -111,4 +110,5 @@ CMakeFiles/fractal_gas_core.dir/src/visit_grid.cpp.o: \
  /usr/include/x86_64-linux-gnu/bits/stdlib.h \
  /usr/include/c++/15/bits/std_abs.h \
  /usr/include/c++/15/pstl/glue_algorithm_defs.h \
- /usr/include/c++/15/pstl/execution_defs.h
+ /usr/include/c++/15/pstl/execution_defs.h /usr/include/c++/15/utility \
+ /usr/include/c++/15/bits/stl_relops.h